refactor(renaming #8): changed the pluginName and identity
