+++ src/main/kotlin/com/damorosodaragona/jenkinsnotifier/CiStatusDebugLog.kt
@@ -1,4 +1,4 @@
-package com.skillab.projector.cistatus
+package com.damorosodaragona.jenkinsnotifier
 
 import com.intellij.openapi.application.PathManager
 import com.intellij.openapi.project.Project
