+++ src/main/kotlin/com/damorosodaragona/jenkinsnotifier/CiStatusEvents.kt
@@ -1,4 +1,4 @@
-package com.skillab.projector.cistatus
+package com.damorosodaragona.jenkinsnotifier
 
 import com.intellij.util.messages.Topic
