+++ src/test/kotlin/com/damorosodaragona/jenkinsnotifier/JenkinsBuildSummaryTest.kt
@@ -1,4 +1,4 @@
-package com.skillab.projector.cistatus
+package com.damorosodaragona.jenkinsnotifier
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
